Do the actual job
