+++ src/backend/distributed/-.d
@@ -1,0 +1,1 @@
+-: sql/citus--7.0-1--7.0-2.sql
