+++ src/backend/distributed/abc
@@ -1,0 +1,1 @@
+-.o:
