+++ src/backend/distributed/master/master_truncate.c
@@ -18,8 +18,8 @@
 #include "distributed/commands/utility_hook.h"
 #include "distributed/master_metadata_utility.h"
 #include "distributed/master_protocol.h"
+#include "distributed/multi_executor.h"
 #include "distributed/multi_join_order.h"
-#include "distributed/multi_router_executor.h"
 #include "distributed/pg_dist_partition.h"
 #include "distributed/resource_lock.h"
 #include "utils/builtins.h"
